refactor: Move storage bg out
